Add custom hook useFilter to handle search functionality
